image handling improvements
improvements to enable more economical image consumption in manhattan.
add a tap for substituting alternate sized chips, nominally expected to be
used to make an intermediate size image from a giant image.  Also added
new logic to graphics to detect images that are being drawn where they
will not be seen (and so don't need to be loaded).